BookingServiceImpl edited

+++ booking-manager-service/src/test/java/service/BookingServiceImplTest.java
@@ -9,7 +9,9 @@ import java.util.Calendar;
 import java.util.Date;
 import org.dozer.DozerBeanMapper;
 import org.junit.Assert;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import static org.mockito.Mockito.when;
 import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
@@ -38,6 +40,7 @@ public class BookingServiceImplTest {
     Date d;
 
     @Autowired
+    @InjectMocks
     BookingService bookingService;
     
     @BeforeMethod
@@ -52,8 +55,10 @@ public class BookingServiceImplTest {
     
     @Test
     public void testAddBooking(){
+	when(bookingDao.getBookingById(booking.getId())).thenReturn(booking);
+	
 	bookingService.addBooking(booking);
-	Assert.assertEquals(bookingDao.getBookingById(booking.getId()).getCheckIn(), d);
+	Assert.assertEquals(bookingService.getBookingById(booking.getId()).getCheckIn(), d);
     }
     
     @Test
@@ -63,22 +68,30 @@ public class BookingServiceImplTest {
 	Date newDate = addDays(d,1);
 	booking.setCheckIn(newDate);
 	bookingDao.updateBooking(booking);
-	Assert.assertEquals(bookingDao.getBookingById(booking.getId()).getCheckIn(), newDate);
+	
+	when(bookingDao.getBookingById(booking.getId()).getCheckIn()).thenReturn(newDate);
+	Assert.assertEquals(bookingService.getBookingById(booking.getId()).getCheckIn(), newDate);
     }
     
     @Test
     public void testDeleteBooking(){
 	bookingDao.addBooking(booking);
-        Assert.assertNotNull(bookingDao.getBookingById(booking.getId()));
+	
+	when(bookingDao.getBookingById(booking.getId())).thenReturn(booking);
+        Assert.assertNotNull(bookingService.getBookingById(booking.getId()));
 
         bookingDao.deleteBooking(booking);
-        Assert.assertNull(bookingDao.getBookingById(booking.getId()));
+	
+	when(bookingDao.getBookingById(booking.getId())).thenReturn(null);
+        Assert.assertNull(bookingService.getBookingById(booking.getId()));
     }
     
     @Test
     public void testGetBookingById(){
-        bookingDao.addBooking(booking);
-        Booking output = bookingDao.getBookingById(booking.getId());
+        bookingService.addBooking(booking);
+        Booking output = bookingService.getBookingById(booking.getId());
+	
+	when(bookingDao.getBookingById(booking.getId())).thenReturn(booking);
         Assert.assertEquals(output.getCheckIn(), booking.getCheckIn());
     }
